Tests now use the number of tuples instead of components in the priority array

+++ src/Reader/DataReader/JsonReader.cpp
@@ -239,7 +239,7 @@ vtkSmartPointer<vtkPolyData> JsonReader::getVtkDataSet(int zoomLevel) {
             Configuration::getInstance().getInteger("dataReader.maximumPriority")
             - (*iterator)->getPriority()
         );*/
-        int priority[1] = {
+        double priority[1] = {
             Configuration::getInstance().getInteger("dataReader.maximumPriority")
             - (*iterator)->getPriority()
         };
